Fixing config and drizzle issues
