+++ config/production.js
@@ -1,9 +1,6 @@
-import { AppConfig } from "@/lib/models/config"
-
-
 const defaultFhirBaseUrl = 'https://identity-matching.fast.hl7.org/fhir';
 
-export const config: Partial<AppConfig> = {
+export const config = {
   env: 'production',
   port: 4000,
   defaultFhirBaseUrl: defaultFhirBaseUrl,
