+++ drizzle.config.ts
@@ -4,7 +4,7 @@ import { appConfig } from '@/config'
 
 export default defineConfig({
   out: './drizzle',
-  schema: './src/db/schema',
+  schema: './src/db/schema.ts',
   dialect: 'sqlite',
   dbCredentials: {
     url: appConfig.database.url,
