+++ src/config.ts
@@ -1,4 +1,4 @@
 import config from 'config';
-import { AppConfig } from "@/lib/models/config"
+import { AppConfig } from "./lib/models/config"
 
 export const appConfig = config.util.toObject() as AppConfig;
